Added team plot comments.
I also ran some of these plots with the data that was partitioned on minutes with no improvement. So, I think it is best if we just show the regression analysis from the partitioning and point out that the regression analysis shows no improvement.  This makes intuitive sense.  This is a good example of a data set which cannot bennefit from partitioning on one variable.  I do think cluster analysis could help by identifying certain "player types."  At this point we do not have time for that, though.

Signed-off-by: BillRowcliffe <ieguitarist@gmail.com>